+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog for NeoFS Node
 
 ### Changed
 - SNs exchange TLS certificates on inter-node connections (#4097)
+- SN no longer adds origin signature to EC requests sent to remote nodes with API >= v2.25 (#4118)
 
 ### Removed